service/controller: merge clusterapi key package with common
After removing obsolete code there weren't a lot of conflicts between
`service/controller/clusterapi/v22/key` and `service/controller/key` so
it is possible to merge the two together providing a flatter namespace.
