--- service/controller/clusterapi/v22/key/error.go
@@ -1,21 +1,0 @@
-package key
-
-import "github.com/giantswarm/microerror"
-
-var notFoundError = &microerror.Error{
-	Kind: "notFoundError",
-}
-
-// IsNotFound asserts notFoundError.
-func IsNotFound(err error) bool {
-	return microerror.Cause(err) == notFoundError
-}
-
-var wrongTypeError = &microerror.Error{
-	Kind: "wrongTypeError",
-}
-
-// IsWrongTypeError asserts wrongTypeError.
-func IsWrongTypeError(err error) bool {
-	return microerror.Cause(err) == wrongTypeError
-}
