+++ service/controller/resource/operatorversions/resource.go
@@ -10,8 +10,8 @@ import (
 	"github.com/giantswarm/micrologger"
 	"github.com/giantswarm/versionbundle"
 
-	"github.com/giantswarm/cluster-operator/service/controller/clusterapi/v22/key"
 	"github.com/giantswarm/cluster-operator/service/controller/controllercontext"
+	"github.com/giantswarm/cluster-operator/service/controller/key"
 )
 
 const (
